% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_accessibility.R
\name{check_palette}
\alias{check_palette}
\title{Check palette accessibility with colorblind simulations}
\usage{
check_palette(name, type = "discrete", n = NULL)
}
\arguments{
\item{name}{Palette name (character)}

\item{type}{Either "discrete" or "continuous" (default "discrete")}

\item{n}{Number of colors for continuous palettes}
}
\description{
Uses \code{colorblindr} to simulate common forms of colorblindness.
}
\examples{
\dontrun{
vangogh::check_palette("StarryNight")
}
}
