% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_palette_scores.R
\name{generate_cvd_scores_data}
\alias{generate_cvd_scores_data}
\title{Generate CVD Scores Data for Package}
\usage{
generate_cvd_scores_data()
}
\value{
A data frame with CVD scores
}
\description{
Internal function to generate the vangogh_cvd_scores dataset.
Run this when updating the package to refresh CVD scores.
}
\keyword{internal}
