% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vangogh_cvd.R
\name{plot_cvd_simulation}
\alias{plot_cvd_simulation}
\title{Plot CVD Simulation for a Palette}
\usage{
plot_cvd_simulation(palette_name, results)
}
\arguments{
\item{palette_name}{Character string with palette name}

\item{results}{List of CVD simulation results from check_vangogh_cvd}
}
\description{
Plot CVD Simulation for a Palette
}
\keyword{internal}
