% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_fill_vangogh}
\alias{scale_fill_vangogh}
\title{#' Setup fill palette for ggplot2}
\usage{
scale_fill_vangogh(
  ...,
  palette = "StarryNight",
  discrete = TRUE,
  alpha = 1,
  reverse = TRUE
)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'vangogh_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\value{
A scale_fill_vangogh function
}
\description{
#' Setup fill palette for ggplot2
}
