% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vangogh.R
\name{theme_vangogh}
\alias{theme_vangogh}
\title{Theme inspired by Van Gogh (variants)}
\usage{
theme_vangogh(
  base_size = 12,
  base_family = "",
  variant = c("classic", "light", "dark", "sketch")
)
}
\arguments{
\item{base_size}{numeric base font size}

\item{base_family}{font family}

\item{variant}{one of "classic", "light", "dark", "sketch"}
}
\description{
Theme inspired by Van Gogh (variants)
}
