% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_accessibility.R
\name{vangogh_colors}
\alias{vangogh_colors}
\title{Return all Van Gogh palettes as a tidy data frame}
\usage{
vangogh_colors(
  n = NULL,
  type = "discrete",
  colorblind = FALSE,
  add_metadata = FALSE
)
}
\arguments{
\item{n}{Number of colors per palette}

\item{type}{"discrete" or "continuous"}

\item{colorblind}{Logical (compatibility)}

\item{add_metadata}{Logical: compute HCL metadata if colorspace available}
}
\description{
Return all Van Gogh palettes as a tidy data frame
}
