% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vangogh_interpolate.R
\name{vangogh_interpolate}
\alias{vangogh_interpolate}
\title{Interpolate a Van Gogh palette}
\usage{
vangogh_interpolate(palette, n)
}
\arguments{
\item{palette}{Vector of hex colors}

\item{n}{Number of colors desired}
}
\description{
Interpolate a Van Gogh palette
}
