% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{vangogh_pal}
\alias{vangogh_pal}
\title{vangogh palette with ramped colours}
\usage{
vangogh_pal(palette = "StarryRhone", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'vangogh_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\value{
Palettes with ramped colors from predefined palettes
}
\description{
vangogh palette with ramped colours
}
\examples{
library(scales)
show_col(vangogh_pal()(10))

filled.contour(volcano, color.palette = vangogh_pal(), asp = 1)
}
