% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_accessibility.R
\name{vangogh_palette_info}
\alias{vangogh_palette_info}
\title{Return palette info as a data frame with optional HCL}
\usage{
vangogh_palette_info(colorblind = FALSE, add_metadata = FALSE)
}
\arguments{
\item{colorblind}{Logical (compatibility)}

\item{add_metadata}{Logical: compute HCL hue/chroma/luminance if colorspace is installed}
}
\description{
Return palette info as a data frame with optional HCL
}
