% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_palette_scores.R
\name{vangogh_palette_info_with_cvd}
\alias{vangogh_palette_info_with_cvd}
\title{Add CVD Information to Palette Documentation}
\usage{
vangogh_palette_info_with_cvd(palette_name)
}
\arguments{
\item{palette_name}{Character string specifying the palette name}
}
\value{
List with palette info and CVD scores
}
\description{
Helper function to add CVD accessibility information to your
existing palette information functions.
}
\examples{
\dontrun{
# Get enhanced palette info with CVD data
info <- vangogh_palette_info_with_cvd("StarryNight")
print(info$cvd_accessibility$summary)
}

}
