% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_accessibility.R
\name{vangogh_suggest}
\alias{vangogh_suggest}
\title{Suggest a palette based on number of colors}
\usage{
vangogh_suggest(n = 5, type = "discrete")
}
\arguments{
\item{n}{Number of colors needed}

\item{type}{"discrete" or "continuous"}
}
\description{
Suggest a palette based on number of colors
}
