% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_vangogh.R
\name{scale_fill_vangogh}
\alias{scale_fill_vangogh}
\title{Scale fill with Van Gogh palettes}
\usage{
scale_fill_vangogh(name, discrete = TRUE, colorblind = FALSE, ...)
}
\arguments{
\item{name}{Palette name}

\item{discrete}{Logical: use discrete scale}

\item{colorblind}{Logical: use colorblind-safe colors}

\item{...}{Additional arguments to ggplot2 scale function}
}
\description{
Scale fill with Van Gogh palettes
}
