% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vangogh.R
\docType{package}
\name{vangogh}
\alias{vangogh}
\title{vangogh}
\description{
list of palettes generated from Vincent van Gogh's paintings
}
\details{
list of palettes generated from Vincent van Gogh's paintings
}
